rename hooks folder to _hooks
